Consolidate partially migrates suites
